# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Pocket Money API
  description: Backend API for managing chores, ledger entries, and settlements in family groups
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.pocket-money.example.com
    description: Production server
...
paths:
...
  /api/v1/groups/{id}/ledger:
    get:
      tags:
        - Ledger
      summary: List ledger entries
      description: Returns ledger entries for a group. Members see only their entries, head sees all.
      operationId: listLedger
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID (UUID)
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          description: Filter by status
          schema:
            type: string
            enum: [approved, pending_approval, rejected]
        - name: user_id
          in: query
          required: false
          description: Filter by user ID (head only)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of ledger entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LedgerResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not a member of this group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Ledger
      summary: Create ledger entry
      description: Creates a new ledger entry. Members create pending entries for themselves, head can create approved entries for any member.
      operationId: createLedger
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID (UUID)
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLedgerRequest'
      responses:
        '201':
          description: Ledger entry created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerResponse'
        '400':
          description: Bad request (validation error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permission denied (e.g., only head can create settlement entries)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/ledger/{id}/approve:
    post:
      tags:
        - Ledger
      summary: Approve ledger entry
      description: Approves a pending ledger entry (head only)
      operationId: approveLedger
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Ledger entry ID (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Entry approved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerResponse'
        '400':
          description: Entry is not pending approval
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Only group head can approve entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/ledger/{id}/reject:
    post:
      tags:
        - Ledger
      summary: Reject ledger entry
      description: Rejects a pending ledger entry (head only)
      operationId: rejectLedger
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Ledger entry ID (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Entry rejected successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LedgerResponse'
        '400':
          description: Entry is not pending approval
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Only group head can reject entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/groups/{id}/pending:
    get:
      tags:
        - Ledger
      summary: List pending entries
      description: Returns all pending ledger entries for a group (head only)
      operationId: listPending
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of pending entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LedgerResponse'
        '400':
          description: Invalid group ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Only group head can view pending entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/groups/{id}/balance:
    get:
      tags:
        - Ledger
      summary: Get group balances
      description: Returns balance for each member in the group
      operationId: getBalance
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Group ID (UUID)
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of member balances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BalanceResponse'
        '400':
          description: Invalid group ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not a member of this group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
...
    CreateLedgerRequest:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
          nullable: true
          description: Target user ID (optional, only head can specify, defaults to self)
        chore_id:
          type: string
          format: uuid
          description: Chore ID
        amount:
          type: number
          format: double
          nullable: true
          description: Custom amount (required only for system chores like Settlement)
      required:
        - chore_id
    LedgerResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Ledger entry ID
        group_id:
          type: string
          format: uuid
          description: Group ID
        user_id:
          type: string
          format: uuid
          description: User who earned the amount
        chore_id:
          type: string
          format: uuid
          description: Chore ID
        amount:
          type: number
          format: double
          description: Amount earned
        status:
          type: string
          enum: [approved, pending_approval, rejected]
          description: Entry status
        created_by_user_id:
          type: string
          format: uuid
          description: User who created the entry
        approved_by_user_id:
          type: string
          format: uuid
          nullable: true
          description: User who approved the entry
        rejected_by_user_id:
          type: string
          format: uuid
          nullable: true
          description: User who rejected the entry
        created_at:
          type: string
          format: date-time
          description: Entry creation timestamp
      required:
        - id
        - group_id
        - user_id
        - chore_id
        - amount
        - status
        - created_by_user_id
        - created_at
    BalanceResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
          description: User ID
        name:
          type: string
          description: User's full name
        balance:
          type: number
          format: double
          description: User's current balance (sum of approved ledger entries minus settlements)
      required:
        - user_id
        - name
        - balance
...